+++ src/main/kotlin/cloud/glitchdev/rfu/config/categories/SeaCreatureConfig.kt
@@ -6,6 +6,7 @@ import cloud.glitchdev.rfu.constants.fishing.RareScDisplayDataType
 import cloud.glitchdev.rfu.constants.fishing.SeaCreatures
 import cloud.glitchdev.rfu.feature.fishing.CatchMessageReplacer
 import cloud.glitchdev.rfu.feature.fishing.RareScPartyMessage
+import cloud.glitchdev.rfu.feature.mob.RareAlert
 import cloud.glitchdev.rfu.gui.window.SeaCreatureEditWindow
 import cloud.glitchdev.rfu.utils.dsl.toExactRegex
 import com.teamresourceful.resourcefulconfig.api.types.options.TranslatableValue
@@ -64,13 +65,13 @@ object SeaCreatureConfig : Category("Sea Creatures") {
 
     var catchMessageTemplate by string("&3&lSEA CREATURE! &eYou caught {article} {style}&l{name}") {
         name = Literal("Catch Message Template")
-        description = Literal("The template for the catch message. Available: {article}, {article_upper}, {name}, {style}, {plural}, {mob}, {mobs}")
+        description = Literal("The template for the catch message. Available: {article}, {article_upper}, {name}, {style}, {color}, {plural}, {mob}, {mobs}")
         condition = { replaceCatchMessages }
     }
 
     var doubleHookCatchMessageTemplate by string("&9&lDOUBLE HOOK! &eYou caught two {style}&l{plural}") {
         name = Literal("Double Hook Message Template")
-        description = Literal("The template for the double hook catch message. Available: {article}, {article_upper}, {name}, {style}, {plural}, {mob}, {mobs}")
+        description = Literal("The template for the double hook catch message. Available: {article}, {article_upper}, {name}, {style}, {color}, {plural}, {mob}, {mobs}")
         condition = { replaceCatchMessages }
     }
 
@@ -92,6 +93,20 @@ object SeaCreatureConfig : Category("Sea Creatures") {
     var detectionAlert by reloadableBoolean(true) {
         name = Literal("Rare Sc Alert")
         description = Literal("Sends an alert whenever a rare SC is found.")
+    }
+
+    var rareScAlertPreset by string("&6&l[&fα&6&l] {color}&l{name} &6&l[&fα&6&l]") {
+        name = Literal("Rare Sc Alert Preset")
+        description = Literal("The template for the rare SC alert. Available: {article}, {article_upper}, {name}, {style}, {color}, {plural}, {mob}, {mobs}")
+        condition = { detectionAlert }
+    }
+
+    init {
+        previewButton(
+            RareAlert::preview,
+            "Preview Alert",
+            "Shows a preview of the rare SC alert."
+        ) { detectionAlert }
     }
 
     var rareScSound by reloadableBoolean(true) {